requirements.txt
asgiref==3.8.1
black==24.4.1
click==8.1.7
colorama==0.4.6
coverage==7.5.0
Django==5.0.4
django-htmx==1.17.3
django-taggit==5.0.1
factory-boy==3.3.0
Faker==24.11.0
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
psycopg==3.1.18
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ruff==0.4.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
